Refactor summary components
and address a couple of other points from PR feedback.
