+++ lib/requests/RequestDataTableColumns.js
@@ -1,0 +1,17 @@
+/**
+ * Columns for `<FcDataTableRequests>`.
+ */
+const RequestDataTableColumns = [
+  { value: 'ID', text: 'ID' },
+  { value: 'data-table-expand', text: 'Location' },
+  { value: 'STUDY_TYPE', text: 'Type' },
+  { value: 'REQUESTER', text: 'Requester' },
+  { value: 'CREATED_AT', text: 'Date Created' },
+  { value: 'ASSIGNED_TO', text: 'Assigned To' },
+  { value: 'DUE_DATE', text: 'Due Date' },
+  { value: 'STATUS', text: 'Status' },
+  { value: 'LAST_EDITED_AT', text: 'Last Updated' },
+  { value: 'ACTIONS', text: '' },
+];
+
+export default RequestDataTableColumns;
